remove ~dirty variable from Camera class. add eCoordinateSystem, eClipRange parameter to Camera::GerProjectionMaterix, Camera::GerViewMaterix, Camera::GerViewProjectionMaterix

+++ Doom3/Source/GameLogic/CustomComponent/AutoRotate.cpp
@@ -7,7 +7,7 @@ void dooms::AutoRotate::InitComponent()
 
 void dooms::AutoRotate::UpdateComponent()
 {
-	GetTransform()->Rotate({ RotateSpeed.x * MainTimer::GetSingleton()->GetDeltaTime(),  RotateSpeed.y * MainTimer::GetSingleton()->GetDeltaTime(), RotateSpeed.z * MainTimer::GetSingleton()->GetDeltaTime() }, eSpace::World);
+	GetTransform()->Rotate({ static_cast<FLOAT32>(RotateSpeed.x * MainTimer::GetSingleton()->GetDeltaTime()), static_cast<FLOAT32>(RotateSpeed.y * MainTimer::GetSingleton()->GetDeltaTime()), static_cast<FLOAT32>(RotateSpeed.z * MainTimer::GetSingleton()->GetDeltaTime()) }, eSpace::World);
 
 	
 	//D_DEBUG_DRAW_3D_LINE(GetTransform()->GetPosition(), GetTransform()->GetPosition() + GetTransform()->forward() * 5, eColor::Red);
